requirements.txt
django
djangorestframework
oauth2client
google-api-python-client
django-cors-headers
